Initial commit of component-ising particle system to allow for better re-use.
